:recycle: Separate file for OperationSet
