+++ Conv/Algorithm/Algorithm.swift
@@ -8,22 +8,6 @@
 
 import Foundation
 
-struct OperationSet {
-    var sectionInsert: [Int] = []
-    var sectionUpdate: [Int] = []
-    var sectionDelete: [Int] = []
-    var sectionMove: [(source: Int, target: Int)] = []
-    
-    var itemInsert: [DifferenciableIndexPath] = []
-    var itemUpdate: [DifferenciableIndexPath] = []
-    var itemDelete: [DifferenciableIndexPath] = []
-    var itemMove: [(source: DifferenciableIndexPath, target: DifferenciableIndexPath)] = []
-    
-    init() {
-        
-    }
-}
-
 struct Diff {
     func diffItem(
         from oldIndexPaths: [DifferenciableIndexPath],
